Allow non-avatar OGB Penetrators to use tps autoconfigure

+++ Scripts/Editor/VF/PlayModeTrigger.cs
@@ -2,10 +2,12 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using BestHTTP.SecureProtocol.Org.BouncyCastle.Math.EC.Multiplier;
 using UnityEditor;
 using UnityEngine;
 using UnityEngine.SceneManagement;
 using VF.Builder;
+using VF.Builder.Exceptions;
 using VF.Inspector;
 using VF.Menu;
 using VF.Model;
@@ -18,6 +20,8 @@ namespace VF {
     public class PlayModeTrigger : IVRCSDKPostprocessAvatarCallback {
         private static double lastRescan = 0;
         private static string AboutToUploadKey = "vrcf_vrcAboutToUpload";
+        private const string tmpDirParent = "Assets/_VRCFury/PlayMode";
+        private static string tmpDir;
         public int callbackOrder => int.MaxValue;
         public void OnPostprocessAvatar() {
             EditorPrefs.SetFloat(AboutToUploadKey, Now());
@@ -50,6 +54,8 @@ namespace VF {
                     var rootObjects = GetRootObjects().ToArray();
                     VRCFPrefabFixer.Fix(rootObjects);
                 }
+
+                tmpDir = null;
             } else if (state == PlayModeStateChange.EnteredPlayMode) {
                 if (problyUploading) {
                     EditorPrefs.DeleteKey(AboutToUploadKey);
@@ -79,11 +85,12 @@ namespace VF {
             if (!Application.isPlaying) return;
             if (!PlayModeMenuItem.Get()) return;
             if (!activeNow) return;
-            
-            var tmpDirParent = $"Assets/_VRCFury/PlayMode";
-            VRCFuryAssetDatabase.DeleteFolder(tmpDirParent);
-            var tmpDir = $"{tmpDirParent}/{DateTime.Now.ToString("yyyyMMdd-HHmmss")}";
-            Directory.CreateDirectory(tmpDir);
+
+            if (tmpDir == null) {
+                VRCFuryAssetDatabase.DeleteFolder(tmpDirParent);
+                tmpDir = $"{tmpDirParent}/{DateTime.Now.ToString("yyyyMMdd-HHmmss")}";
+                Directory.CreateDirectory(tmpDir);
+            }
 
             var builder = new VRCFuryBuilder();
             var oneChanged = false;
@@ -104,13 +111,19 @@ namespace VF {
                     oneChanged = true;
                 }
                 foreach (var o in root.GetComponentsInChildren<OGBOrifice>(true)) {
+                    if (!o.gameObject.activeInHierarchy) continue;
                     if (ContainsAnyPrefabs(o.gameObject)) continue;
-                    OGBOrificeEditor.Bake(o, onlySenders: true);
+                    VRCFExceptionUtils.ErrorDialogBoundary(() => {
+                        OGBOrificeEditor.Bake(o, onlySenders: true);
+                    });
                     Object.DestroyImmediate(o);
                 }
                 foreach (var o in root.GetComponentsInChildren<OGBPenetrator>(true)) {
+                    if (!o.gameObject.activeInHierarchy) continue;
                     if (ContainsAnyPrefabs(o.gameObject)) continue;
-                    OGBPenetratorEditor.Bake(o, onlySenders: true, tmpDir: tmpDir);
+                    VRCFExceptionUtils.ErrorDialogBoundary(() => {
+                        OGBPenetratorEditor.Bake(o, onlySenders: true, tmpDir: tmpDir);
+                    });
                     Object.DestroyImmediate(o);
                 }
             }
